+++ port/linux/package/pikascript/pikascript-lib/pika_cjson/pika_cjson_cJSON.c
@@ -95,7 +95,7 @@ int pika_cjson_cJSON_getType(PikaObj* self) {
     return item->type;
 }
 
-float pika_cjson_cJSON_getValueDouble(PikaObj* self) {
+double pika_cjson_cJSON_getValueDouble(PikaObj* self) {
     cJSON* item = obj_getPtr(self, "item");
     return item->valuedouble;
 }
